simplify server startup; break tests into local and remote modules
