+++ Cargo.toml
@@ -41,4 +41,3 @@ tokio = { version = "0.2.0", features = [ "full" ] }
 serial_test = "0.4"
 serial_test_derive = "0.4"
 warp = "0.2.2"
-reqwest =  {version = "0.10.4", features = ["blocking"]}
