+++ tests/local.rs
@@ -4,177 +4,9 @@ extern crate fantoccini;
 extern crate futures_util;
 
 use fantoccini::{error, Client};
-use std::net::{Ipv4Addr, TcpListener, SocketAddr};
-use std::path::PathBuf;
-use warp::Filter;
-use futures_core::Future;
-use fantoccini::error::CmdError;
-use warp::test::request;
 
+mod common;
 
-async fn select_client_type(s: &str) -> Result<Client, error::NewSessionError> {
-    match s {
-        "firefox" => {
-            let mut caps = serde_json::map::Map::new();
-            let opts = serde_json::json!({ "args": ["--headless"] });
-            caps.insert("moz:firefoxOptions".to_string(), opts.clone());
-            Client::with_capabilities("http://localhost:4444", caps).await
-        },
-        "chrome" => {
-            let mut caps = serde_json::map::Map::new();
-            let opts = serde_json::json!({
-                "args": ["--headless", "--disable-gpu", "--no-sandbox", "--disable-dev-shm-usage"],
-                "binary":
-                    if std::path::Path::new("/usr/bin/chromium-browser").exists() {
-                        // on Ubuntu, it's called chromium-browser
-                        "/usr/bin/chromium-browser"
-                    } else if std::path::Path::new("/Applications/Google Chrome.app/Contents/MacOS/Google Chrome").exists() {
-                        // macOS
-                        "/Applications/Google Chrome.app/Contents/MacOS/Google Chrome"
-                    } else {
-                        // elsewhere, it's just called chromium
-                        "/usr/bin/chromium"
-                    }
-            });
-            caps.insert("goog:chromeOptions".to_string(), opts.clone());
-
-            Client::with_capabilities("http://localhost:9515", caps).await
-        },
-        browser => unimplemented!("unsupported browser backend {}", browser),
-    }
-}
-
-fn handle_test_error(res: Result<Result<(), CmdError>, Box<dyn std::any::Any + Send>>) -> bool{
-    match res {
-        Ok(Ok(_)) => true,
-        Ok(Err(e)) => {
-            eprintln!("test future failed to resolve: {:?}", e);
-            false
-        }
-        Err(e) => {
-            if let Some(e) = e.downcast_ref::<error::CmdError>() {
-                eprintln!("test future panicked: {:?}", e);
-            } else if let Some(e) = e.downcast_ref::<error::NewSessionError>() {
-                eprintln!("test future panicked: {:?}", e);
-            } else {
-                eprintln!("test future panicked; an assertion probably failed");
-            }
-            false
-        }
-    }
-}
-
-macro_rules! local_tester {
-    ($f:ident, $endpoint:expr) => {{
-        use std::sync::{Arc, Mutex};
-        use std::thread;
-
-        let port: u16 = setup_server();
-        let c = select_client_type($endpoint);
-
-        // we'll need the session_id from the thread
-        // NOTE: even if it panics, so can't just return it
-        let session_id = Arc::new(Mutex::new(None));
-
-        // run test in its own thread to catch panics
-        let sid = session_id.clone();
-        let res = thread::spawn(move || {
-        let mut rt = tokio::runtime::Builder::new().enable_all().basic_scheduler().build().unwrap();
-            let mut c = rt.block_on(c).expect("failed to construct test client");
-            *sid.lock().unwrap() = rt.block_on(c.session_id()).unwrap();
-            let x = rt.block_on($f(c, port));
-            drop(rt);
-            x
-        })
-        .join();
-        let success = handle_test_error(res);
-        assert!(success);
-    }}
-}
-
-
-macro_rules! tester {
-    // Ident should identify an async fn that takes just a Client.
-    ($f:ident, $endpoint:expr) => {{
-        use std::sync::{Arc, Mutex};
-        use std::thread;
-
-        let c = select_client_type($endpoint);
-
-        // we'll need the session_id from the thread
-        // NOTE: even if it panics, so can't just return it
-        let session_id = Arc::new(Mutex::new(None));
-
-        // run test in its own thread to catch panics
-        let sid = session_id.clone();
-        let res = thread::spawn(move || {
-            let mut rt = tokio::runtime::Builder::new().enable_all().basic_scheduler().build().unwrap();
-            let mut c = rt.block_on(c).expect("failed to construct test client");
-            *sid.lock().unwrap() = rt.block_on(c.session_id()).unwrap();
-            let x = rt.block_on($f(c));
-            drop(rt);
-            x
-        });
-        let success = handle_test_error(res);
-        assert!(success);
-    }};
-
-}
-
-fn setup_server() -> u16 {
-    let (tx, rx) = std::sync::mpsc::channel();
-
-     std::thread::spawn(move || {
-         let mut rt = tokio::runtime::Builder::new().enable_all().basic_scheduler().build().unwrap();
-         let _ = rt.block_on(async {
-             let socket_addr = TcpListener::bind("127.0.0.1:0")
-                .unwrap()
-                .local_addr()
-                .unwrap();
-             let (socket_addr, server) = start_server(socket_addr);
-             tx.send(socket_addr.port()).expect("To be able to send");
-             server
-        });
-    });
-
-    let port = rx.recv().expect("to get port");
-    println!("got port: {}", port);
-
-    let url = format!("http://localhost:{}/sample_page.html", port);
-    let mut count = 0;
-    loop {
-        match reqwest::blocking::get(&url) {
-            Ok(..) => break,
-            e => {
-                println!("error: {:?}", e);
-                count += 1;
-                if count > 100 {
-                    break;
-                }
-
-                std::thread::sleep(std::time::Duration::from_millis(100));
-            }
-        }
-    }
-    port
-}
-
-/// Starts the fileserver
-fn start_server(socket_addr: SocketAddr) -> (SocketAddr, impl Future<Output = ()> + 'static) {
-    const ASSETS_DIR: &str = "tests/test_html";
-    let assets_dir: PathBuf = PathBuf::from(ASSETS_DIR);
-    let routes = fileserver(assets_dir);
-    warp::serve(routes).bind_ephemeral(socket_addr)
-}
-
-/// Serves files under this directory.
-fn fileserver(
-    assets_dir: PathBuf,
-) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-    warp::get()
-        .and(warp::fs::dir(assets_dir))
-        .and(warp::path::end())
-}
 
 fn sample_page_url(port: u16) -> String {
     format!("http://localhost:{}/sample_page.html", port)
